drupal/imagemagick 5.0.0-rc1, drupal/image_effects 5.0.0-rc1, drupal/sophron 3.0.0

+++ web/modules/contrib/image_effects/src/Plugin/ImageToolkit/Operation/gd/TextToWrapper.php
@@ -171,7 +171,7 @@ class TextToWrapper extends GDImageToolkitOperationBase {
       $current_y += $line_height;
       switch ($arguments['text_align']) {
         case 'center':
-          $x_offset = round($x_delta / 2);
+          $x_offset = (int) round($x_delta / 2);
           break;
 
         case 'right':
